catching up

+++ code/user_list_indexer/src/com/turning_leaf_technologies/reindexer/UserListIndexer.java
@@ -1,5 +1,6 @@
 package com.turning_leaf_technologies.reindexer;
 
+import com.turning_leaf_technologies.encryption.EncryptionUtils;
 import com.turning_leaf_technologies.indexing.IndexingUtils;
 import com.turning_leaf_technologies.indexing.Scope;
 import org.apache.logging.log4j.Logger;
@@ -35,13 +36,15 @@ class UserListIndexer {
 	private HashSet<Long> usersThatCanShareLists = new HashSet<>();
 	private SolrClient openArchivesServer;
 	private PreparedStatement getListDisplayNameAndAuthorStmt;
+	private final String serverName;
 
-	UserListIndexer(Ini configIni, Connection dbConn, Logger logger){
+	UserListIndexer(String serverName, Ini configIni, Connection dbConn, Logger logger){
+		this.serverName = serverName;
 		this.dbConn = dbConn;
 		this.logger = logger;
 		//Load a list of all list publishers
 		try {
-			PreparedStatement listPublishersStmt = dbConn.prepareStatement("SELECT userId FROM `user_roles` INNER JOIN roles on user_roles.roleId = roles.roleId where name = 'listPublisher'");
+			PreparedStatement listPublishersStmt = dbConn.prepareStatement("SELECT userId FROM user_roles INNER JOIN roles on user_roles.roleId = roles.roleId inner join role_permissions on role_permissions.roleId = roles.roleId where permissionId = (select id from permissions where name = 'Include Lists In Search Results')");
 			ResultSet listPublishersRS = listPublishersStmt.executeQuery();
 			while (listPublishersRS.next()){
 				usersThatCanShareLists.add(listPublishersRS.getLong(1));
@@ -66,7 +69,6 @@ class UserListIndexer {
 		groupedWorkServer = groupedWorkHttpBuilder.build();
 		HttpSolrClient.Builder openArchivesHttpBuilder = new HttpSolrClient.Builder("http://localhost:" + solrPort + "/solr/open_archives");
 		openArchivesServer = openArchivesHttpBuilder.build();
-
 
 		scopes = IndexingUtils.loadScopes(dbConn, logger);
 	}
@@ -104,13 +106,11 @@ class UserListIndexer {
 				updateServer.deleteByQuery("recordtype:list");
 				//Get a list of all public lists
 				numListsStmt = dbConn.prepareStatement("select count(id) as numLists from user_list WHERE deleted = 0 AND public = 1 and searchable = 1");
-				//noinspection SpellCheckingInspection
-				listsStmt = dbConn.prepareStatement("SELECT user_list.id as id, deleted, public, searchable, title, description, user_list.created, dateUpdated, username, firstname, lastname, displayName, homeLocationId, user_id from user_list INNER JOIN user on user_id = user.id WHERE public = 1 AND deleted = 0");
+				listsStmt = dbConn.prepareStatement("SELECT user_list.id as id, deleted, public, searchable, title, description, user_list.created, dateUpdated, username, firstname, lastname, displayName, homeLocationId, user_id from user_list INNER JOIN user on user_id = user.id WHERE public = 1 AND searchable = 1 AND deleted = 0");
 			}else{
 				//Get a list of all lists that are were changed since the last update
 				//Have to process all lists because one could have been deleted, made private, or made non searchable.
 				numListsStmt = dbConn.prepareStatement("select count(id) as numLists from user_list");
-				//noinspection SpellCheckingInspection
 				listsStmt = dbConn.prepareStatement("SELECT user_list.id as id, deleted, public, searchable, title, description, user_list.created, dateUpdated, username, firstname, lastname, displayName, homeLocationId, user_id from user_list INNER JOIN user on user_id = user.id WHERE dateUpdated > ?");
 				listsStmt.setLong(1, lastReindexTime);
 			}
@@ -138,7 +138,7 @@ class UserListIndexer {
 			}
 
 			while (allPublicListsRS.next()){
-				if (updateSolrForList(updateServer, getTitlesForListStmt, allPublicListsRS, lastReindexTime, logEntry)){
+				if (updateSolrForList(fullReindex, updateServer, getTitlesForListStmt, allPublicListsRS, lastReindexTime, logEntry)){
 					numListsIndexed++;
 				}
 				numListsProcessed++;
@@ -154,7 +154,7 @@ class UserListIndexer {
 		return numListsProcessed;
 	}
 
-	private boolean updateSolrForList(ConcurrentUpdateSolrClient updateServer, PreparedStatement getTitlesForListStmt, ResultSet allPublicListsRS, long lastReindexTime, ListIndexingLogEntry logEntry) throws SQLException, SolrServerException, IOException {
+	private boolean updateSolrForList(boolean fullReindex, ConcurrentUpdateSolrClient updateServer, PreparedStatement getTitlesForListStmt, ResultSet allPublicListsRS, long lastReindexTime, ListIndexingLogEntry logEntry) throws SQLException, SolrServerException, IOException {
 		UserListSolr userListSolr = new UserListSolr(this);
 		long listId = allPublicListsRS.getLong("id");
 
@@ -163,7 +163,7 @@ class UserListIndexer {
 		int isSearchable = allPublicListsRS.getInt("searchable");
 		long userId = allPublicListsRS.getLong("user_id");
 		boolean indexed = false;
-		if (deleted == 1 || isPublic == 0 || isSearchable == 0){
+		if (!fullReindex && (deleted == 1 || isPublic == 0 || isSearchable == 0)){
 			updateServer.deleteByQuery("id:" + listId);
 			logEntry.incDeleted();
 		}else{
@@ -174,114 +174,118 @@ class UserListIndexer {
 			long created = allPublicListsRS.getLong("created");
 			userListSolr.setCreated(created);
 
-			String displayName = allPublicListsRS.getString("displayName");
-			//noinspection SpellCheckingInspection
-			String firstName = allPublicListsRS.getString("firstname");
-			//noinspection SpellCheckingInspection
-			String lastName = allPublicListsRS.getString("lastname");
-			String userName = allPublicListsRS.getString("username");
+			try {
+				String displayName = EncryptionUtils.decryptString(allPublicListsRS.getString("displayName"), serverName, logEntry);
+				String firstName = EncryptionUtils.decryptString(allPublicListsRS.getString("firstname"), serverName, logEntry);
+				String lastName = EncryptionUtils.decryptString(allPublicListsRS.getString("lastname"), serverName, logEntry);
+				String userName = allPublicListsRS.getString("username");
 
-			if (userName.equalsIgnoreCase("nyt_user")) {
-				userListSolr.setOwnerCanShareListsInSearchResults(true);
-			}else{
-				userListSolr.setOwnerCanShareListsInSearchResults(usersThatCanShareLists.contains(userId));
-			}
-			if (displayName != null && displayName.length() > 0){
-				userListSolr.setAuthor(displayName);
-			}else{
-				if (firstName == null) firstName = "";
-				if (lastName == null) lastName = "";
-				String firstNameFirstChar = "";
-				if (firstName.length() > 0){
-					firstNameFirstChar = firstName.charAt(0) + ". ";
-				}
-				userListSolr.setAuthor(firstNameFirstChar + lastName);
-			}
-
-			long patronHomeLibrary = allPublicListsRS.getLong("homeLocationId");
-			if (librariesByHomeLocation.containsKey(patronHomeLibrary)){
-				userListSolr.setOwningLibrary(librariesByHomeLocation.get(patronHomeLibrary));
-			} else {
-				//Don't know the owning library for some reason
-				userListSolr.setOwningLibrary(-1);
-			}
-
-			//Don't know the owning location
-			userListSolr.setOwningLocation(locationCodesByHomeLocation.getOrDefault(patronHomeLibrary, ""));
-
-			//Get information about all of the list titles.
-			getTitlesForListStmt.setLong(1, listId);
-			ResultSet allTitlesRS = getTitlesForListStmt.executeQuery();
-			while (allTitlesRS.next()) {
-				String source = allTitlesRS.getString("source");
-				String sourceId = allTitlesRS.getString("sourceId");
-				if (!allTitlesRS.wasNull()){
-					if (sourceId.length() > 0 && source.equals("GroupedWork")) {
-						// Skip archive object Ids
-						SolrQuery query = new SolrQuery();
-						query.setQuery("id:" + sourceId);
-						query.setFields("title_display", "author_display");
-
-						try {
-							QueryResponse response = groupedWorkServer.query(query);
-							SolrDocumentList results = response.getResults();
-							//Should only ever get one response
-							if (results.size() >= 1) {
-								SolrDocument curWork = results.get(0);
-								userListSolr.addListTitle("grouped_work", sourceId, curWork.getFieldValue("title_display"), curWork.getFieldValue("author_display"));
-							}
-						} catch (Exception e) {
-							logger.error("Error loading information about title " + sourceId);
-						}
-					}else if (source.equals("OpenArchives")){
-						// Skip archive object Ids
-						SolrQuery query = new SolrQuery();
-						query.setQuery("id:" + sourceId);
-						query.setFields("title", "creator");
-
-						try {
-							QueryResponse response = openArchivesServer.query(query);
-							SolrDocumentList results = response.getResults();
-							//Should only ever get one response
-							if (results.size() >= 1) {
-								SolrDocument curWork = results.get(0);
-								userListSolr.addListTitle("open_archives", sourceId, curWork.getFieldValue("title"), curWork.getFieldValue("creator"));
-							}
-						} catch (Exception e) {
-							logger.error("Error loading information about title " + sourceId);
-						}
-					}else if (source.equals("Lists")){
-						getListDisplayNameAndAuthorStmt.setString(1, sourceId);
-						ResultSet listDisplayNameAndAuthorRS = getListDisplayNameAndAuthorStmt.executeQuery();
-						if (listDisplayNameAndAuthorRS.next()){
-							userListSolr.addListTitle("lists", sourceId, listDisplayNameAndAuthorRS.getString("title"), listDisplayNameAndAuthorRS.getString("displayName"));
-						}
-						listDisplayNameAndAuthorRS.close();
-					}else{
-						logEntry.incErrors("Unhandled source " + source);
-					}
-					//TODO: Handle other types of objects within a User List
-					//people, etc.
-				}
-			}
-			if (userListSolr.getNumTitles() >= 3) {
-				// Index in the solr catalog
-				SolrInputDocument document = userListSolr.getSolrDocument();
-				if (document != null){
-					updateServer.add(userListSolr.getSolrDocument());
-					if (created > lastReindexTime){
-						logEntry.incAdded();
-					}else{
-						logEntry.incUpdated();
-					}
-					indexed = true;
+				if (userName.equalsIgnoreCase("nyt_user")) {
+					userListSolr.setOwnerCanShareListsInSearchResults(true);
 				}else{
+					userListSolr.setOwnerCanShareListsInSearchResults(usersThatCanShareLists.contains(userId));
+				}
+				if (displayName != null && displayName.length() > 0){
+					userListSolr.setAuthor(displayName);
+				}else{
+					if (firstName == null) firstName = "";
+					if (lastName == null) lastName = "";
+					String firstNameFirstChar = "";
+					if (firstName.length() > 0){
+						firstNameFirstChar = firstName.charAt(0) + ". ";
+					}
+					userListSolr.setAuthor(firstNameFirstChar + lastName);
+				}
+
+				long patronHomeLibrary = allPublicListsRS.getLong("homeLocationId");
+				if (librariesByHomeLocation.containsKey(patronHomeLibrary)){
+					userListSolr.setOwningLibrary(librariesByHomeLocation.get(patronHomeLibrary));
+				} else {
+					//Don't know the owning library for some reason, most likely this is an admin user.
+					userListSolr.setOwningLibrary(-1);
+				}
+
+				//Don't know the owning location
+				userListSolr.setOwningLocation(locationCodesByHomeLocation.getOrDefault(patronHomeLibrary, ""));
+
+				//Get information about all of the list titles.
+				getTitlesForListStmt.setLong(1, listId);
+				ResultSet allTitlesRS = getTitlesForListStmt.executeQuery();
+				while (allTitlesRS.next()) {
+					String source = allTitlesRS.getString("source");
+					String sourceId = allTitlesRS.getString("sourceId");
+					if (!allTitlesRS.wasNull()){
+						if (sourceId.length() > 0 && source.equals("GroupedWork")) {
+							// Skip archive object Ids
+							SolrQuery query = new SolrQuery();
+							query.setQuery("id:" + sourceId);
+							query.setFields("title_display", "author_display");
+
+							try {
+								QueryResponse response = groupedWorkServer.query(query);
+								SolrDocumentList results = response.getResults();
+								//Should only ever get one response
+								if (results.size() >= 1) {
+									SolrDocument curWork = results.get(0);
+									userListSolr.addListTitle("grouped_work", sourceId, curWork.getFieldValue("title_display"), curWork.getFieldValue("author_display"));
+								}
+							} catch (Exception e) {
+								logger.error("Error loading information about title " + sourceId);
+							}
+						}else if (source.equals("OpenArchives")){
+							// Skip archive object Ids
+							SolrQuery query = new SolrQuery();
+							query.setQuery("id:" + sourceId);
+							query.setFields("title", "creator");
+
+							try {
+								QueryResponse response = openArchivesServer.query(query);
+								SolrDocumentList results = response.getResults();
+								//Should only ever get one response
+								if (results.size() >= 1) {
+									SolrDocument curWork = results.get(0);
+									userListSolr.addListTitle("open_archives", sourceId, curWork.getFieldValue("title"), curWork.getFieldValue("creator"));
+								}
+							} catch (Exception e) {
+								logger.error("Error loading information about title " + sourceId);
+							}
+						}else if (source.equals("Lists")){
+							getListDisplayNameAndAuthorStmt.setString(1, sourceId);
+							ResultSet listDisplayNameAndAuthorRS = getListDisplayNameAndAuthorStmt.executeQuery();
+							if (listDisplayNameAndAuthorRS.next()){
+								String decryptedName = EncryptionUtils.decryptString(listDisplayNameAndAuthorRS.getString("displayName"), serverName, logEntry);
+								userListSolr.addListTitle("lists", sourceId, listDisplayNameAndAuthorRS.getString("title"), decryptedName);
+							}
+							listDisplayNameAndAuthorRS.close();
+						}else{
+							logEntry.incErrors("Unhandled source " + source);
+						}
+						//TODO: Handle other types of objects within a User List
+						//people, etc.
+					}
+				}
+				if (userListSolr.getNumTitles() >= 3) {
+					// Index in the solr catalog
+					SolrInputDocument document = userListSolr.getSolrDocument();
+					if (document != null){
+						updateServer.add(document);
+						if (created > lastReindexTime){
+							logEntry.incAdded();
+						}else{
+							logEntry.incUpdated();
+						}
+						indexed = true;
+					}else{
+						updateServer.deleteByQuery("id:" + listId);
+						logEntry.incDeleted();
+					}
+				} else {
 					updateServer.deleteByQuery("id:" + listId);
 					logEntry.incDeleted();
 				}
-			} else {
-				updateServer.deleteByQuery("id:" + listId);
-				logEntry.incDeleted();
+			}catch (Exception e){
+				logEntry.addNote("Could not decrypt user information for " + listId + " - " + e.toString());
+				logEntry.incSkipped();
 			}
 		}
